--- src/test/java/com/fpmislata/daw1/projectedaw1/mock/persistance/dao/EscriuDaoMock.java
@@ -1,28 +1,0 @@
-package com.fpmislata.daw1.projectedaw1.mock.persistance.dao;
-
-import com.fpmislata.daw1.projectedaw1.data.LlibreData;
-import com.fpmislata.daw1.projectedaw1.domain.entity.Autor;
-import com.fpmislata.daw1.projectedaw1.domain.entity.Llibre;
-import com.fpmislata.daw1.projectedaw1.persistance.dao.EscriuDao;
-
-import java.util.List;
-
-public class EscriuDaoMock implements EscriuDao {
-
-    private final List<Llibre> llibreList = LlibreData.LLIBRE_LIST;
-    @Override
-    public List<Autor> findAutorsByLlibreIsbn(String isbn) {
-        return null;
-    }
-
-    @Override
-    public List<Llibre> findLlibresByAutorId(int idAutor) {
-        if (idAutor == 1) {
-            return List.of(llibreList.get(1), llibreList.get(2));
-        } else if (idAutor == 2) {
-            return List.of(llibreList.get(2));
-        } else {
-            return List.of();
-        }
-    }
-}
